refactor(cache): consolidate caching utilities
add `createCache` wrapper for queries
