+++ src/app/[locale]/(city)/[cityId]/(other)/(tabs)/people/page.tsx
@@ -1,7 +1,7 @@
 import { notFound } from "next/navigation";
 import { isUserAuthorizedToEdit } from "@/lib/auth";
 import CityPeople from "@/components/cities/CityPeople";
-import { getPartiesForCityCached, getPeopleForCityCached, getAdministrativeBodiesForCityCached } from "@/lib/cachedData";
+import { getPartiesForCityCached, getPeopleForCityCached, getAdministrativeBodiesForCityCached } from "@/lib/cache";
 
 export default async function PeoplePage({
     params: { cityId }
