+++ src/lib/db/meetings.ts
@@ -79,7 +79,6 @@ export async function getCouncilMeeting(cityId: string, id: string): Promise<Cou
 }
 
 export async function getCouncilMeetingsForCity(cityId: string, { includeUnreleased }: { includeUnreleased: boolean } = { includeUnreleased: false }): Promise<CouncilMeetingWithAdminBodyAndSubjects[]> {
-    console.log(`getCouncilMeetingsForCity: ${cityId} (includeUnreleased: ${includeUnreleased})`);
 
     try {
         const meetings = await prisma.councilMeeting.findMany({
@@ -125,7 +124,6 @@ export async function toggleMeetingRelease(cityId: string, id: string, released:
 }
 
 export async function getCouncilMeetingsCountForCity(cityId: string): Promise<number> {
-    console.log(`getCouncilMeetingsCountForCity: ${cityId}`);
     try {
         const count = await prisma.councilMeeting.count({
             where: {
